clean up and test extensions.model
- make res-unet compatible with modified base
- move ResBlock from layers.building_blocks.py to model.res_unet and
hide it
- implement test for res-unet
- move content of layers.prefab.py to layers.convolutional_blocks.py
