+++ inferno/extensions/model/__init__.py
@@ -1,1 +1,2 @@
-from .unet import ResBlockUNet, UNet
+from .unet import UNet, UNetBase
+from .res_unet import ResBlockUNet
