+++ inferno/extensions/model/unet.py
@@ -1,13 +1,11 @@
 import torch
 import torch.nn as nn
 from ..layers.identity import Identity
-from ..layers.convolutional_blocks import ResBlock
 from ..layers.convolutional import ConvELU2D, ConvELU3D, Conv2D, Conv3D
 from ...utils.math_utils import max_allowed_ds_steps
-from ...utils.python_utils import require_dict_kwagrs
 
 
-__all__ = ['UNetBase', 'ResBlockUNet']
+__all__ = ['UNetBase', 'UNet', 'ResBlockUNet']
 _all = __all__
 
 
@@ -160,7 +158,7 @@ class UNetBase(nn.Module):
         current_in_channels = self.in_channels * self.gain**self.depth
 
         for i in range(self.depth):
-            # the number of out channels (are we in the last block?)
+            # the number of out channels
             out_channels = current_in_channels // self.gain
 
             # if not residual we concat which needs twice as many channels
@@ -315,10 +313,11 @@ class UNetBase(nn.Module):
             return x
 
 
+# TODO implement function to load a pretrained unet
 class UNet(UNetBase):
     """
-    Default 2d / 3d U-Net implementation following
-    TODO cite Ronneberger U-Net papers.
+    Default 2d / 3d U-Net implementation following:
+    https://arxiv.org/abs/1505.04597
     """
     def __init__(self, in_channels, out_channels, dim,
                  depth=4, initial_features=64, gain=2,
@@ -370,58 +369,3 @@ class UNet(UNetBase):
             conv = nn.Sequential(self.default_conv(in_channels, out_channels, 3),
                                  self.default_conv(out_channels, out_channels, 3))
         return conv, False
-
-
-# TODO rename ResidualBlockUNet
-# TODO test
-# or ResidualUNet
-class ResBlockUNet(UNetBase):
-    """TODO.
-
-        ACCC
-
-    Attributes:
-        activated (TYPE): Description
-        dim (TYPE): Description
-        res_block_kwargs (TYPE): Description
-        side_out_parts (TYPE): Description
-        unet_kwargs (TYPE): Description
-    """
-    def __init__(self, in_channels, dim, out_channels, unet_kwargs=None,
-                 res_block_kwargs=None, activated=True,
-                 side_out_parts=None):
-
-        self.dim = dim
-        self.unet_kwargs = require_dict_kwagrs(unet_kwargs, "unet_kwargs must be a dict or None")
-        self.res_block_kwargs = require_dict_kwagrs(res_block_kwargs,
-                                                    "res_block_kwargs must be a dict or None")
-        self.activated = activated
-        if isinstance(side_out_parts, str):
-            self.side_out_parts = set([side_out_parts])
-        elif isinstance(side_out_parts, (tuple,list)):
-            self.side_out_parts = set(side_out_parts)
-        else:
-            self.side_out_parts = set()
-
-        super(ResBlockUNet, self).__init__(in_channels=in_channels,
-                                           dim=dim,
-                                           out_channels=out_channels,
-                                           **self.unet_kwargs)
-
-    def conv_op_factory(self, in_channels, out_channels, part, index):
-
-        # is this the very last convolutional block?
-        very_last = (part == 'up' and index + 1 == self.depth)
-
-        # should the residual block be activated?
-        activated = not very_last or self.activated
-
-        # should the output be part of the overall
-        # return-list in the forward pass of the UNet
-        use_as_output = part in self.side_out_parts
-
-        # residual block used within the UNet
-        # TODO rename to ResidualBlock
-        return ResBlock(in_channels=in_channels, out_channels=out_channels,
-                             dim=self.dim, activated=activated,
-                             **self.res_block_kwargs), use_as_output
